Qfsm - Member variables documented according to Qfsm source codes. - Member variables reordered to match Qfsm source codes. - Several data types of member variables fixed. - Several items are optional. - Arrow type and transition type are enums. - Member variables renamed to match Java conventions and to be better usable.

+++ src/net/sourceforge/anotherfsm/qfsm/QfsmInitialTransition.java
@@ -24,8 +24,15 @@ package net.sourceforge.anotherfsm.qfsm;
  * @author Michal Turek
  */
 public class QfsmInitialTransition {
-	public int xpos;
-	public int ypos;
-	public int endx;
-	public int endy;
+	/** The X coordinate of the position. */
+	public double xpos;
+
+	/** The Y coordinate of the position. */
+	public double ypos;
+
+	/** The X coordinate of the end position. */
+	public double endx;
+
+	/** The X coordinate of the end position. */
+	public double endy;
 }
